improve script, rename entry file
