+++ src/App.tsx
@@ -1,7 +1,7 @@
 import * as React from "react";
 import { EVENTS } from "./model/events";
 // import arSA from "date-fns/locale/ar-SA";
-import { Scheduler } from "./lib/Scheduler";
+import { Scheduler } from "./lib";
 
 const App = () => {
   const [events, setEvents] = React.useState(EVENTS);
