+++ webpack.config.js
@@ -6,10 +6,10 @@ const nodeExternals = require("webpack-node-externals");
 module.exports = {
   mode: "production",
   target: "node",
-  entry: path.resolve(__dirname, "src", "lib", "Scheduler.tsx"),
+  entry: path.resolve(__dirname, "src", "lib", "index.tsx"),
   output: {
     path: path.resolve(__dirname, "dist"),
-    filename: "Scheduler.js",
+    filename: "index.js",
     libraryTarget: "commonjs2",
   },
   module: {
